+++ test/lib/SignatureHelper.sol
@@ -28,9 +28,11 @@ contract SignatureHelper is Test {
 
     function structHash(QuarkWallet.QuarkOperation memory op) public view returns (bytes32) {
         bytes memory encodedArray;
-        for (uint i = 0; i < op.scriptSources.length;) {
+        for (uint256 i = 0; i < op.scriptSources.length;) {
             encodedArray = abi.encodePacked(encodedArray, keccak256(op.scriptSources[i]));
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
 
         return keccak256(
